backend/requirements.txt
Django==5.2.4
djangorestframework==3.16.0
asgiref==3.9.1
sqlparse==0.5.3
django-cors-headers

python-dotenv
gunicorn
psycopg2-binary

pip install djangorestframework djoser djangorestframework-simplejwt
